fix array key warnings on rendering
This change basically ignores the warnings by using null (which is the
default on a warning anyway). This is not a proper fix for the problem
(which is mostly related to $INFO not always being set). However
properly fixing the issue would require major refactoring of how
multiple tables are handled and how compatibility with the include
plugin is established.

This partly changes how 041c5927655e6a845b75a2503b8c2978615646a9 tried
to fix similar errors